Do no overwrite systemd service file

+++ scripts/install.sh
@@ -80,8 +80,11 @@ configure_systemd() {
     status "Adding current user to ollama group..."
     $SUDO usermod -a -G ollama $(whoami)
 
-    status "Creating ollama systemd service..."
-    cat <<EOF | $SUDO tee /etc/systemd/system/ollama.service >/dev/null
+    if [ -f "/etc/systemd/system/ollama.service" ]; then
+    status "Ollama systemd service exists..."
+    else
+        status "Creating ollama systemd service..."
+        cat <<EOF | $SUDO tee /etc/systemd/system/ollama.service >/dev/null
 [Unit]
 Description=Ollama Service
 After=network-online.target
@@ -97,6 +100,7 @@ Environment="PATH=$PATH"
 [Install]
 WantedBy=default.target
 EOF
+    fi
     SYSTEMCTL_RUNNING="$(systemctl is-system-running || true)"
     case $SYSTEMCTL_RUNNING in
         running|degraded)
